Moved attaching library patterns from bindDelegation to validate
Moved attaching patterns from library to elements from bindDelegation to validate.  Added the ability to combine custom patterns set for elements with the use of the pattern library.  The validation uses both the custom pattern and the pattern pulled from the library if combineClassName in settings is true.  Otherwise, it defaults to the custom pattern if it exists, followed by the one found in the pattern library if it doesn't.

We needed this because in bindDelegation it was just overwriting custom validation that was in place with the pattern attribute.  This way you can still use both validations.